remove abstract class
